// back/src/bot/enums/services.enum.ts
export enum ServicesE {
    companyRegistration = "Регистрация компании",
    companyTransfer = "Перевод компании под наше обслуживание",
    financialStatements = "Подгатовка/Подача финансовой отчетности",
    additionalDocuments = "Дополнительные документы",
    vatNumber = "VAT номер",
    helpAccountIBAN = "Помощь в открытии счета IBAN",
    helpAccountMerchant = "Помощь в открытии merchant счета",
    residencePermit = "Получение ВНЖ",
}